+++ plugins/orchestrator-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 ### Dependencies
 
+* **@janus-idp/cli:** upgraded to 1.15.1
+
+### Dependencies
+
+* **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.18.2
+
+### Dependencies
+
 * **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.18.1
 
 ### Dependencies